+++ src/main/java/vn/edu/fpt/petworldplatform/dto/HealthCheckContextDTO.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 public class HealthCheckContextDTO {
     private Integer appointmentId;
     private String appointmentCode;
-    private Long petId;
+    private Integer petId;
     private String petName;
     private String status;
     private String customerName;
